requirements.txt
SQLAlchemy==1.4.46
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
PyMySQL==1.1.0
WTForms==3.0.1
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
click==8.1.4
werkzeug==2.3.7